requirements.txt
streamlit
openai
langchain
newspaper3k
faiss-cpu
python-dotenv
